Fix breaking indicator list when common indicators are not instantiated
Suppose there is only one indicator in the plan, and this indicator is
linked to (i.e., an instantiation of) a common indicator A. Further
assume that A has another common indicator B among its effects, but B
has no instantiation in the plan. Since A has an effect, the old code
will not consider A a root in the dependency graph. However, since the
dependency graph's only node is A, there will be no root.

This will cause the `path` attribute to be missing from A because
'expandPaths` never visits A. The code will try to access `path` and
fail.