Become compatible with the new theme builds
There are two changes in this commit:

- The new themes contains relative paths in variables now, we need to
use the getThemeStaticURL to convert them to absolute paths now.

- The new theme infrastructure does not compile all the available svg icons
into one combined svg to be references with the svg use
attribute. Instead, we include all the available icons inside a hidden
DOM element in the UI code.

The latter change also includes a new way to use Icons -- instead
of saying <Icon name="icon-name" /> we can now say <Icon.IconName />
but also the old usage is supported for more dynamic use cases.

This was implemented because we need a TypeScript enum for the
available icons since they need to be available at runtime so
we can include all of them inside the DOM, and once we have
the Enum, it's easy to use it to make the icon usage simpler
and still type-safe.