Fix bug in Action List Page text search regex handling
Introduced in 06e9a542452e8190ef390f6c06b88eb6683d76b8 - after having
escaped the user provided search string for use as a regular
expression, the identifier search broke since it was not implemented
as a regular expression search but simple string matching. For example
dashes in identifiers like AB-CD-1 where not matched because they are
a special character in regex and had been escaped.