Fix problem with _next/static.* file requests ending up in app code
The way nextjs routing works is that it checks if a static file exists
in the filesystem and if it does not, it will move on to routing
through the application. This results in missing static files
causing errors in app code.

Fix this problem by marking the static file requests specially
in the middleware and exiting early from app code if the request
is thus marked. We need to modify both the app layout and the
middleware, because inside the middleware we do not yet know
if the static request will be succesful. Only unsuccesful
requests end up in the app code.