Use correct term for actions in action page breadcrumb
The code made an attempt to use the correct term before, but there was a
clash between two identical translation keys, defined in different
locale files. Previously only the one that was not context-aware ended
up being the translation that was used.

Rename the other key to avoid the keys overriding each other.