Fix continuous action filter
Previously the filter that makes the action list only show continuous
actions did not work properly: if clicked once, it would work, but after
turning it off again still only continuous actions remained on the list.
By removing the url query parameter when the filter is toggled off, the
filter works as expected.